backend/requirements.txt
# ==========================
# Django + REST
# ==========================
Django==5.1.7
djangorestframework==3.16.1
django-cors-headers
django-eventstream

# ==========================
# Celery + Redis
# ==========================
celery==5.5.3
redis==7.0.1

# ==========================
# LLM + API Client
# ==========================
openai==2.6.1
tiktoken==0.12.0
httpx==0.28.1

# ==========================
# Utils
# ==========================
python-dotenv==1.0.1
django-environ==0.4.5
requests==2.31.0
certifi==2024.8.30

hgtk

# ==========================
# Docs
# ==========================
drf-yasg==1.21.11

# ==========================
# Async Tasks
# ==========================
django-celery-beat==2.8.1

# ==========================
# WSGI
# ==========================
gunicorn==22.0.0
